step2, 테스트 코드 분리 - CalculatorTest - OperatorTest - OperatorInputTest
